// src/app/model/chat.ts
export class Chat{
    from: string;
    to: string;
    type: string;
    message: string;
    picture: any;
    dateTime: any;
    isMessageScene:boolean;
    petReport:{};
    fromUser?: string;
    pet_report_card?: {};
    image?: any;
    toUser?: string;
    systemGenerated?: any;
    createdate?: any;
    booking_user?: string;
}